interactive init
